+++ lib/domain/acoustics.dart
@@ -1,0 +1,18 @@
+/// Quantity types, units and constants commonly encountered in acoustics.
+library acoustics;
+
+export 'package:quantity/quantity.dart'
+    show
+        // Quantity classes and units
+        Frequency,
+        FrequencyUnits,
+        Level,
+        LevelUnits,
+        SoundIntensityLevel,
+        SoundPressureLevel,
+        Time,
+        TimeUnits,
+        Pressure,
+        PressureUnits,
+        Power,
+        PowerUnits;