+++ lib/domain/atomic_nuclear.dart
@@ -1,0 +1,52 @@
+/// Quantity types, units and constants commonly encountered in the fields of atomic, molecular, and nuclear physics.
+library atomic_nuclear;
+
+export 'package:quantity/quantity.dart'
+    show
+        // Quantity classes and units
+        Mass,
+        MassUnits,
+        Length,
+        LengthUnits,
+        Charge,
+        ChargeUnits,
+        Energy,
+        EnergyUnits,
+        Area,
+        AreaUnits,
+        Activity,
+        ActivityUnits,
+        AbsorbedDose,
+        AbsorbedDoseUnits,
+        DoseEquivalent,
+        DoseEquivalentUnits,
+
+        // Particle mass constants
+        alphaParticleMass,
+        atomicMass,
+        electronMass,
+        protonMass,
+        neutronMass,
+        tauMass,
+        deuteronMass,
+        helionMass,
+        muonMass,
+
+        // Other fundamental / particle constants
+        elementaryCharge,
+        planckConstant,
+        hBar,
+        fineStructureConstant,
+        bohrRadius,
+        comptonWavelength,
+        tauComptonWavelength,
+        classicalElectronRadius,
+
+        // Units
+        barns,
+        electronVolts;
+
+export 'package:quantity/domain/universal.dart' show rydberg;
+
+export 'package:quantity/domain/thermodynamic.dart'
+    show avogadro, faradayConstant, loschmidtStdAtm;