+++ lib/domain/chemistry.dart
@@ -1,0 +1,27 @@
+/// Quantity types, units and constants commonly encountered in the field of chemistry.
+library chemistry;
+
+export 'package:quantity/quantity.dart'
+    show
+        // Quantity classes and units
+        AmountOfSubstance,
+        AmountOfSubstanceUnits,
+        Concentration,
+        ConcentrationUnits,
+        MolarEnergy,
+        MolarEnergyUnits,
+        MolarEntropy,
+        MolarEntropyUnits,
+        MassDensity,
+        MassDensityUnits,
+        Volume,
+        VolumeUnits,
+
+        // Particle constants
+        atomicMass,
+
+        // Units
+        moles;
+
+export 'package:quantity/domain/thermodynamic.dart'
+    show avogadro, gasConstantMolar, faradayConstant, loschmidtStdAtm;